fix: fix --version option
