+++ .tool-versions
@@ -1,0 +1,1 @@
+yarn 1.22.22
