+++ jest.config.js
@@ -1,4 +1,5 @@
-// We cannot employ vitest because tests must call require() dynamically.
-module.exports = {
+// We cannot employ vitest because test must call require() dynamically.
+const config = {
   testEnvironment: 'node',
 };
+export default config;
