+++ package.json
@@ -10,7 +10,8 @@
   "repository": "github:WillBooster/gen-i18n-ts",
   "license": "Apache-2.0",
   "author": "WillBooster Inc.",
-  "bin": "./bin/index.cjs",
+  "type": "module",
+  "bin": "./bin/index.js",
   "files": [
     "bin/",
     "dist/"
@@ -20,13 +21,13 @@
     "cleanup": "yarn format && yarn lint-fix",
     "format": "sort-package-json && yarn prettify",
     "postinstall": "husky || true",
-    "lint": "eslint --color \"./{scripts,src,tests}/**/*.{cjs,cts,js,jsx,mjs,mts,ts,tsx}\"",
+    "lint": "eslint --color \"./{scripts,src,test}/**/*.{cjs,cts,js,jsx,mjs,mts,ts,tsx}\"",
     "lint-fix": "yarn lint --fix",
     "prepack": "pinst --disable",
     "postpack": "pinst --enable",
     "prepare": "husky || true",
     "prettify": "prettier --cache --color --write \"**/{.*/,}*.{cjs,css,cts,htm,html,js,json,json5,jsonc,jsx,md,mjs,mts,scss,ts,tsx,vue,yaml,yml}\" \"!**/test-fixtures/**\"",
-    "start": "build-ts run src/start.ts --",
+    "start": "build-ts run src/cli.ts --",
     "start-prod": "yarn build && yarn gen-i18n-ts",
     "test": "yarn jest",
     "typecheck": "tsc --noEmit --Pretty"
