--- src/start.ts
@@ -1,4 +1,0 @@
-import { cli } from './index';
-
-// eslint-disable-next-line unicorn/prefer-top-level-await
-void cli();
